+++ packages/validator-bonds-cli-institutional/src/commands/manage/configureBond.ts
@@ -18,12 +18,14 @@ export function installConfigureBond(program: Command) {
         withToken,
         bondAuthority,
         cpmpe,
+        maxStakeWanted,
       }: {
         voteAccount?: Promise<PublicKey>
         authority?: Promise<WalletInterface | PublicKey>
         withToken: boolean
         bondAuthority?: Promise<PublicKey>
         cpmpe?: BN
+        maxStakeWanted?: BN
       },
     ) => {
       await manageConfigureBond({
@@ -34,7 +36,7 @@ export function installConfigureBond(program: Command) {
         withToken,
         newBondAuthority: await bondAuthority,
         cpmpe,
-        maxStakeWanted: undefined,
+        maxStakeWanted,
       })
     },
   )