+++ packages/validator-bonds-cli/__tests__/test-validator/show.spec.ts
@@ -286,6 +286,7 @@ describe('Show command using CLI', () => {
         voteAccount: voteAccount.toBase58(),
         authority: bondAuthority.publicKey.toBase58(),
         costPerMillePerEpoch: '222 lamports',
+        maxStakeWanted: '2000 SOLs',
       },
     }
     const expectedDataFundingSingleItem = {
@@ -544,6 +545,7 @@ describe('Show command using CLI', () => {
         voteAccount: voteAccount,
         authority: bondAuthority.publicKey,
         costPerMillePerEpoch: '1 lamport',
+        maxStakeWanted: '0 SOL',
       },
     }
     const voteAccountShow = await loadTestingVoteAccount(